requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
google-generativeai==0.7.2
pydantic==2.9.2
python-multipart==0.0.9
pymupdf==1.24.9